move PartitionBuilder to separate file; build sycl tests with icpx
